refactor(cli): move RemoteResult render impl into its own module
Addresses Copilot review on #177.

Break the terminal -> remote module cycle: keep 'impl ResultRender for RemoteResult' in remote/benchmark.rs (the type's module) so terminal::render no longer depends on a CLI command. RemoteResult and its helpers tighten back to private, and remote::benchmark returns to a private module.

Also render a neutral 'Benchmark complete (log analysis disabled)' block when logs were not collected, so a run no longer ends with no output instead of faking a pass/warn verdict.
