+++ crates/cli/src/remote.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-pub(crate) mod benchmark;
+mod benchmark;
 pub mod protocol;
 mod testbed;
